refactor(queue): cleanup, extract common funcs
