+++ src/queue/check.ts
@@ -1,16 +1,18 @@
-import { getConfig } from "../config";
-import { walk, statAsync } from "../fs/async";
-import Scene from "../types/scene";
-import Image from "../types/image";
-import { basename } from "path";
-import * as logger from "../logger";
-import { extractLabels, extractActors, extractScenes } from "../extractor";
-import Jimp from "jimp";
 import ora = require("ora");
-import { indexImages } from "../search/image";
+
+import { getConfig } from "../config";
 import { imageCollection, sceneCollection } from "../database";
-import { fileIsExcluded } from "../types/utility";
-import { imageWithPathExists } from "./utility";
+import { statAsync, walk } from "../fs/async";
+import * as logger from "../logger";
+import { indexImages } from "../search/image";
+import Image from "../types/image";
+import Scene from "../types/scene";
+import {
+  imageWithPathExists,
+  isImportableImage,
+  processImage,
+} from "./image/utility";
+import { isImportableVideo } from "./video/utility";
 
 export async function checkVideoFolders() {
   const config = getConfig();
@@ -27,10 +29,7 @@ export async function checkVideoFolders() {
 
     await walk(folder, [".mp4", ".webm"], async (path) => {
       loader.text = `Scanned ${++numFiles} videos`;
-      if (
-        basename(path).startsWith(".") ||
-        fileIsExcluded(config.EXCLUDE_FILES, path)
-      ) {
+      if (!isImportableVideo) {
         logger.log(`Ignoring file ${path}`);
       } else {
         logger.log(`Found matching file ${path}`);
@@ -60,53 +59,14 @@ export async function checkVideoFolders() {
   );
 }
 
-async function processImage(imagePath: string, readImage = true) {
-  try {
-    const imageName = basename(imagePath);
-    const image = new Image(imageName);
-    image.path = imagePath;
-
-    if (readImage) {
-      const jimpImage = await Jimp.read(imagePath);
-      image.meta.dimensions.width = jimpImage.bitmap.width;
-      image.meta.dimensions.height = jimpImage.bitmap.height;
-      image.hash = jimpImage.hash();
-    }
-
-    // Extract scene
-    const extractedScenes = await extractScenes(imagePath);
-    logger.log(`Found ${extractedScenes.length} scenes in image path.`);
-    image.scene = extractedScenes[0] || null;
-
-    // Extract actors
-    const extractedActors = await extractActors(imagePath);
-    logger.log(`Found ${extractedActors.length} actors in image path.`);
-    await Image.setActors(image, [...new Set(extractedActors)]);
-
-    // Extract labels
-    const extractedLabels = await extractLabels(imagePath);
-    logger.log(`Found ${extractedLabels.length} labels in image path.`);
-    await Image.setLabels(image, [...new Set(extractedLabels)]);
-
-    // await database.insert(database.store.images, image);
-    await imageCollection.upsert(image._id, image);
-    await indexImages([image]);
-    logger.success(`Image '${imageName}' done.`);
-  } catch (error) {
-    logger.error(error);
-    logger.error(`Failed to add image '${imagePath}'.`);
-  }
-}
-
-export async function checkImageFolders() {
+export async function checkImageFolders(readImages) {
   const config = getConfig();
 
   logger.log("Checking image folders...");
 
   let numAddedImages = 0;
 
-  if (!config.READ_IMAGES_ON_IMPORT)
-    logger.warn("Reading images on import is disabled.");
+  if (!readImages) logger.warn("Reading images on import is disabled.");
 
   if (config.EXCLUDE_FILES.length)
     logger.log(`Will ignore files: ${config.EXCLUDE_FILES}.`);
@@ -118,14 +78,10 @@ export async function checkImageFolders() {
 
     await walk(folder, [".jpg", ".jpeg", ".png", ".gif"], async (path) => {
       loader.text = `Scanned ${++numFiles} images`;
-      if (
-        basename(path).startsWith(".") ||
-        fileIsExcluded(config.EXCLUDE_FILES, path)
-      )
-        return;
+      if (!isImportableImage(path)) return;
 
       if (!(await imageWithPathExists(path))) {
-        await processImage(path, config.READ_IMAGES_ON_IMPORT);
+        await processImage(path, readImages);
         numAddedImages++;
         logger.log(`Added image '${path}'.`);
       } else {
